LC-equal-sum-grid-partition-i [Runtime: N/A; Memory: N/A]

+++ enumeration/equal-sum-grid-partition-i.cpp
@@ -12,7 +12,7 @@ public:
             }
         }
 
-        int topSum = 0;
+        long long topSum = 0;
         for(int i=0; i<m-1; i++){
             for(int j=0; j<n; j++){
                 topSum += grid[i][j];
